+++ utils/create_mnist_img.py
